backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "cryptography>=46.0.3",
    "fastapi[standard]>=0.115.0",
    "pymysql>=1.1.2",
    "sqlalchemy>=2.0.45",
    "google-generativeai>=0.8.0",
    "httpx>=0.28.0",
    "opensearch-py>=3.1.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.10",
    "pytest>=8.0",
    "pytest-cov>=6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py314"
line-length = 120

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "B008",  # function call in default argument (FastAPI Depends pattern)
]

[tool.ruff.lint.per-file-ignores]
"alembic/env.py" = ["E402"]  # module level import not at top (alembic structure)

[tool.mypy]
python_version = "3.14"
mypy_path = "."
explicit_package_bases = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

# サードパーティライブラリの型チェック
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "entity.*"
disable_error_code = ["misc"]  # SQLAlchemy Base subclass issue
